Type of timeout was changed to long to be consistent with parent class. Tests for Connection.join and Connection.waitAndJoin(url, timeout) were added (only for resource manager).
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/scheduling/trunk@11356 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ src/resource-manager/src/org/ow2/proactive/resourcemanager/frontend/RMConnection.java
@@ -100,7 +100,7 @@ public class RMConnection extends Connection<RMAuthentication> {
      * zero is interpreted as an infinite timeout. The connection will then
      * block until established or an error occurs.
      */
-    public static RMAuthentication waitAndJoin(String url, int timeout) throws RMException {
+    public static RMAuthentication waitAndJoin(String url, long timeout) throws RMException {
         try {
             return getInstance().waitAndConnect(normalize(url), timeout);
         } catch (Exception e) {
